+++ src/RcsbBoard/RcsbDisplay/RcsbBlockDisplay.ts
@@ -12,7 +12,7 @@ import {
 	RcsbFvTrackDataElementGapInterface,
 	RcsbFvTrackDataElementInterface
 } from "../../RcsbDataManager/RcsbDataManager";
-import * as classes from "../scss/RcsbBoard.module.scss";
+import classes from "../scss/RcsbBoard.module.scss";
 import {RcsbD3Constants} from "../RcsbD3/RcsbD3Constants";
 
 export class RcsbBlockDisplay extends RcsbAbstractDisplay {