+++ src/RcsbBoard/RcsbDisplay/RcsbSequenceDisplay.ts
@@ -1,9 +1,8 @@
 import {RcsbAbstractDisplay} from "./RcsbAbstractDisplay";
 import {Selection,BaseType} from "d3-selection";
-import {RcsbDisplayInterface} from "./RcsbDisplayInterface";
 import {scaleLinear, ScaleLinear} from "d3-scale";
 import {LocationViewInterface} from "../RcsbBoard";
-import * as classes from "../scss/RcsbBoard.module.scss";
+import classes from "../scss/RcsbBoard.module.scss";
 import {
     PlotSequenceInterface,
     MoveSequenceInterface,